refactor(cli): split the daemon-client wire types and the nested subcommand enums out
Both `commands.rs` (413 LOC) and `daemon_client/mod.rs` (362 LOC) sit on
the ratchet, and both are about to gain fields: a `--profile` flag group
on three commands, and a `default_profile` on the proxy start body. Make
the room first so the commits that need it stay about what they change.

`subcommands.rs` takes `ChatCommand`, `ProxyCommand` and `DaemonCommand`.
They are the part of `commands.rs` with no coupling to the top level —
none of them names a flag group — so the split costs no shared context.
Re-exported from `commands`, so `dispatch.rs` keeps naming them
`crate::commands::ChatCommand` and the move is invisible to every caller.

`daemon_client/wire.rs` takes the four request/response bodies. The module
it leaves behind owns the *connection* — finding the daemon, launching it,
checking its identity for switch and version skew — which is a different
concern from the shapes travelling over it. The filename deliberately
mirrors `gglib_axum::handlers::proxy::wire`: the two are halves of one
contract and now say so from the tree.

Pure moves. `commands.rs` 413 to 320, `daemon_client/mod.rs` 362 to 325,
no test changes, `cargo test -p gglib-cli` identical before and after.

Full pre-commit green.
