+++ crates/gglib-cli/src/lib.rs
@@ -49,6 +49,7 @@ pub(crate) mod presentation;
 pub(crate) mod reasoning_args;
 pub(crate) mod sampling_params;
 pub(crate) mod shared_args;
+pub(crate) mod subcommands;
 pub(crate) mod utils;
 
 // Re-export primary types for convenient access
